多项BUG修复和逻辑修改

+++ src/main/java/org/haic/often/util/ThreadUtil.java
@@ -34,7 +34,7 @@ public class ThreadUtil {
 	 *
 	 * @param millis 线程等待时间 (毫秒)
 	 */
-	public static void waitThread(int millis) {
+	public static void waitThread(long millis) {
 		try { // 程序等待
 			TimeUnit.MILLISECONDS.sleep(millis);
 		} catch (InterruptedException e) {
